Moved player movement into new MovingSprite super
